+++ nomad-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- refactor: Change DB methods to use explicit generics instead of `impl Trait`
 - require `Common: std::fmt::Display`
 - refactor: Add IRSA credentials to client instantiation
 - implement `Encode` and `Decode` for `bool`